+++ entities/ticket.go
@@ -12,8 +12,8 @@ type Ticket struct {
 	Title                  autotask.Optional[string]    `json:"title,omitzero"`
 	Description            autotask.Optional[string]    `json:"description,omitzero"`
 	TicketNumber           autotask.Optional[string]    `json:"ticketNumber,omitzero"`
-	Status                 autotask.Optional[int]       `json:"status,omitzero"`
-	Priority               autotask.Optional[int]       `json:"priority,omitzero"`
+	Status                 autotask.Optional[int64]       `json:"status,omitzero"`
+	Priority               autotask.Optional[int64]       `json:"priority,omitzero"`
 	QueueID                autotask.Optional[int64]     `json:"queueID,omitzero"`
 	CompanyID              autotask.Optional[int64]     `json:"companyID,omitzero"`
 	CompanyLocationID      autotask.Optional[int64]     `json:"companyLocationID,omitzero"`
@@ -26,11 +26,11 @@ type Ticket struct {
 	CreateDate             autotask.Optional[time.Time] `json:"createDate,omitzero"`
 	LastActivityDate       autotask.Optional[time.Time] `json:"lastActivityDate,omitzero"`
 	CompletedDate          autotask.Optional[time.Time] `json:"completedDate,omitzero"`
-	TicketType             autotask.Optional[int]       `json:"ticketType,omitzero"`
-	IssueType              autotask.Optional[int]       `json:"issueType,omitzero"`
-	SubIssueType           autotask.Optional[int]       `json:"subIssueType,omitzero"`
-	TicketCategory         autotask.Optional[int]       `json:"ticketCategory,omitzero"`
-	Source                 autotask.Optional[int]       `json:"source,omitzero"`
+	TicketType             autotask.Optional[int64]       `json:"ticketType,omitzero"`
+	IssueType              autotask.Optional[int64]       `json:"issueType,omitzero"`
+	SubIssueType           autotask.Optional[int64]       `json:"subIssueType,omitzero"`
+	TicketCategory         autotask.Optional[int64]       `json:"ticketCategory,omitzero"`
+	Source                 autotask.Optional[int64]       `json:"source,omitzero"`
 	BillingCodeID          autotask.Optional[int64]     `json:"billingCodeID,omitzero"`
 	EstimatedHours         autotask.Optional[float64]   `json:"estimatedHours,omitzero"`
 	ExternalID             autotask.Optional[string]    `json:"externalID,omitzero"`
